backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
sentence-transformers
pgvector
beautifulsoup4
requests
python-dotenv